Adds "runtime" package to the cobalt_client package.
This is needed in order to have the Cobalt FIDL service
loaded by bootstrap. Previously we had the "runtime_config"
package but that is useless without also having the
"runtime" package.

Note that this change is only to make the cobalt_client tests
succeed. In real builds of Fuchsia we always have the
"runtime" package anyway.

Change-Id: I03c80ff54a7785be29302e1fa2b734e4308418be