Remove ability to add things to /boot from Fuchsia build
This removes the mechanism for adding parts of the Fuchsia build to
the /boot partition, which is no longer needed with the
devmgr.require-system flag to tell the device manager to hold off on
binding tagged drivers until /system is mounted so it can choose the
libmsd_vendor.so driver when appropriate. This allows simplifying the
boot manifest processing to only select the relevant tags and means
that libc++, libc++abi and libunwind can be mounted into /system
instead of /boot.

Change-Id: I5db3514232b54f131db35530819eb71e3a923a11